Move jam following to hud
