+++ features/main/hud/src/main/java/com/vgleadsheets/features/main/hud/HudState.kt
@@ -4,6 +4,7 @@ import com.airbnb.mvrx.Async
 import com.airbnb.mvrx.MavericksState
 import com.airbnb.mvrx.Uninitialized
 import com.vgleadsheets.features.main.hud.search.SearchContent
+import com.vgleadsheets.model.Jam
 import com.vgleadsheets.model.Part
 import com.vgleadsheets.model.Song
 import com.vgleadsheets.perf.tracking.common.FrameTimeStats
@@ -27,6 +28,7 @@ data class HudState(
     val loadTimeLists: Map<PerfSpec, ScreenLoadStatus>? = null,
     val frameTimeStatsMap: Map<PerfSpec, FrameTimeStats>? = null,
     val invalidateStatsMap: Map<PerfSpec, InvalidateStats>? = null,
+    val activeJam: Jam? = null,
     val perfViewState: PerfViewState = PerfViewState(),
     val updateTime: Async<Long> = Uninitialized,
     val digest: Async<Unit> = Uninitialized
