+++ features/main/jam/src/main/java/com/vgleadsheets/features/main/jam/Clicks.kt
@@ -1,14 +1,16 @@
 package com.vgleadsheets.features.main.jam
 
 import com.vgleadsheets.FragmentRouter
+import com.vgleadsheets.features.main.hud.HudViewModel
 import com.vgleadsheets.features.main.list.ListItemClicks
 
 class Clicks(
     private val router: FragmentRouter,
+    private val hudViewModel: HudViewModel,
     private val viewModel: JamViewModel
 ) : ListItemClicks {
     fun follow(jamId: Long) {
-        router.showJamViewer(jamId)
+        hudViewModel.followJam(jamId)
     }
 
     fun refresh() {
