+++ features/main/jam/src/main/java/com/vgleadsheets/features/main/jam/JamFragment.kt
@@ -45,6 +45,7 @@ class JamFragment : BetterListFragment<JamContent, JamState>() {
         baseImageUrl,
         Clicks(
             getFragmentRouter(),
+            hudViewModel,
             viewModel
         ),
         perfTracker,
